+++ examples/security-grpc-server/src/main/java/net/devh/boot/grpc/examples/security/server/SecurityGrpcServerApplication.java
@@ -15,31 +15,21 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.devh.boot.grpc.examples.cloud.server;
+package net.devh.boot.grpc.examples.security.server;
 
-import org.springframework.security.access.annotation.Secured;
-
-import io.grpc.stub.StreamObserver;
-import net.devh.boot.grpc.examples.lib.HelloReply;
-import net.devh.boot.grpc.examples.lib.HelloRequest;
-import net.devh.boot.grpc.examples.lib.SimpleGrpc;
-import net.devh.boot.grpc.server.service.GrpcService;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
- * An example service that checks the user's authentication.
+ * An example server application that demonstrates basic auth security.
  *
  * @author Daniel Theuke (daniel.theuke@heuboe.de)
  */
-@GrpcService
-public class GrpcServerService extends SimpleGrpc.SimpleImplBase {
+@SpringBootApplication
+public class SecurityGrpcServerApplication {
 
-    // A grpc method that requests the user to be authenticated and have the role "ROLE_GREET".
-    @Override
-    @Secured("ROLE_GREET")
-    public void sayHello(final HelloRequest req, final StreamObserver<HelloReply> responseObserver) {
-        final HelloReply reply = HelloReply.newBuilder().setMessage("Hello ==> " + req.getName()).build();
-        responseObserver.onNext(reply);
-        responseObserver.onCompleted();
+    public static void main(final String[] args) {
+        SpringApplication.run(SecurityGrpcServerApplication.class, args);
     }
 
 }
